Move drivetrain-specific types into respective trajectory generator headers
